pyproject.toml
[project]
name = "gitreport"
version = "0.1.0"
description = "Generate Excel reports from git commit history"
readme = "README.md"
requires-python = ">=3.12"
license = "MIT"
authors = [{ name = "Titusz Pan", email = "tp@py7.de" }]
dependencies = [
    "click>=8.0",
    "gitpython>=3.0",
    "tablib[xlsx]>=3.0",
]

[project.scripts]
gitreport = "main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
only-include = ["main.py"]
